backend/requirements.txt
# FRA Atlas DSS Backend Dependencies
# Core Framework
flask==2.3.3
flask-restful==0.3.10
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-jwt-extended==4.5.3

# Database & Spatial
psycopg2-binary==2.9.7
geoalchemy2==0.14.1
sqlalchemy==2.0.21

# Geospatial Processing
geopandas==0.14.0
shapely==2.0.1
pyproj==3.6.1
rasterio==1.3.8
fiona==1.9.4

# Scientific Computing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.2

# Satellite Data Processing
sentinelsat==1.2.1
rasterio==1.3.8
opencv-python==4.8.0.76

# Machine Learning
scikit-learn==1.3.0
tensorflow==2.13.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.1
redis==4.6.0
gunicorn==21.2.0

# Development & Testing
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0

# Documentation
flask-restx==1.1.0